disdvi.c: tweak
git-svn-id: svn://tug.org/texlive/trunk/Build/source@63196 c570f23f-e606-0410-a88d-b1316a301751

+++ texk/dvi2tty/dvi2tty-src/disdvi.c
@@ -690,7 +690,7 @@ unsigned long num(int size)
 
     pc += size;
     for (i = size; i > 0; i--)
-        x = (x << 8) + (unsigned) getc(dvifp);
+        x = (x << 8) + (unsigned long) getc(dvifp);
 
     return x;
